Single common task to install zabbix-server
Install the zabbix-server-{{ zabbix_server_database }} from a
single common task. Sure, the repo setup is quite different based
on os_family, but the package is practically the same, though there
are some things to pay attention to.

RedHat can pin the minor version, and has a toggle to disable repo.
Debian can't pin minor version, but has install_recommends.

We use the common package module, which is just a wrapper around
apt/yum, and use this construction;

  user_supplied_var | default(_calculated_var | default(omit))

to send additional parameters to the respective modules. You can
kinda read it like this;

If the user wants to override something, use that.
If it's not defined, try a calculated variable based on os_family.
If there's no calculated var, that means the option is not valid
for the package module, so omit this parameter.

I think it's good practice to prefix variables, that are calculated
or not meant for user-overrides with '_'.