Install MySQL dependencies from mysql task
It's cleaner, less conditional checking.

The RedHat playbook already had a variable for the packages, we
give it a slighly more accurate name, and create a similar one for
Debian.

python2 has been sunset sometime back in January 2020, so we're
gonna assume it's always python3. I'm not sure why there's a
special case for debian-buster, where we opt for mariadb-client
instead of default-mysql-client, as the latter package is available
and seems to work. So we're dropping the special buster case.

Arguably, we should switch from mysqldb to pymysql for debian, but
that deserves its separate commit.