Install PostgreSQL dependencies from postgresql task
It's cleaner, less conditional checking.

Similar to the mysql dependencies cleanup.

python2 has been sunset sometime back in January 2020, so we're
gonna assume it's always python3. There's a condition that checks
for zabbix_server_database_creation before installing the python
dependencies, but it's only checking this for RedHat, which leaves
systems inconsistent, so we're gonna drop that conditional.

The tags on this task are quite useful, and should be on the mysql
dependencies aswell. Also, zabbix_server_install_database_client
exists on the postgres side of things, but only for RedHat. Bring
this feature to all. Again, for consistency.