Refactor PostgreSQL schema import
We take advantage of the community.postgres modules, and do a query
for 'mandatory' in 'dbversion', as is done in the mysql tasks. If
the query fails, we're going to assume the database has not been
populated yet, and populate it in the rescue section.

The query and schema import happens from the zabbix server, it
serves as a good test of verifying permissions previously created.