Simplify logic for dbhost_run_install (mysql)
It was difficult to see what this code was supposed to achieve. The
most important variable zabbix_server_dbhost_run_install was hidden
behind two conditions of delegated_dbhost. But here we try to bring
it back to view.

if zabbix_server_dbhost_run_install is true, we want the code to be
delegated_to the zabbix_server_dbhost, otherwise we want to run it
from the zabbix_server.

However, we have a funny situation where zabbix_server_dbhost is
by default localhost, which means we'd still be on the
zabbix_server. We store this calculation in delegated_dbhost.

  if (dbhost == localhost) then zabbix_server else dbhost

We also have an additional variable zabbix_server_real_dbhost,
which we try applying first and if not, we default to whatever
zabbix_server_dbhost_run_install wants.

We move the create db and create user into a block, so conditions
and vars can be shared across both.