Refactor MySQL schema import
We check for existing dbversion in a block, if the query fails, we
create the database in a rescue section. There are a few things to
note however.

The check and schema creation happens from zabbix_server. This is
to verify that our dbuser can access the dbname from zabbix_server
and has all the permissions needed (after creation) to alter tables
between zabbix version upgrades. (So we're not copying schema to
other places, running with escalated privileges anymore)

For the schema creation we actually need some extra privileges, so
we read what they currently are, set them to what we need, import
our schema, and revert the extra privileges back to their original
state. The reverting happens in the `always:` section, so if any
other task should fail in schema creation (`rescue:` section) we
always revert the privileges before failing.