Make molecule container names unique-ish
The self-hosted GitHub actions-runner does not scale with a single
runner on a "powerful" machine.

But we can create multiple users on a single machine, and have each
user run an actions-runner. Problem is they all share the same
docker daemon. And when multiple runners try starting up a container
with the same name as an existing one, the jobs fail.

This is an attempt to make the container names unique enough for
each job, and allows us to scale out actions-runners on a single
machine.