Install selinux packages regardless
Use the generic `package` module for package installs.

Drop when conditions and install selinux packages and dependencies
regardless. Not sure why we install python3-libsemanage for EL8
before we install it again in the package list below, move all
those packages up to the first install.

Unsure why the python library for policycoreutils was only for EL7,
it is needed for EL8 aswell.