Remove useless selinux checking code
It was probably used to determine wheter to include selinux.yml.
But having it moved inside selinux.yml, this code does nothing.

Currently the prerequisite for selinux.yml being included is
zabbix_server_selinux manually set to true by the user/deployer.

Then from inside selinux.yml, we check if selinux is enforcing or
permissive. If it is, we use set_fact to set zabbix_server_selinux
to true (which it already has to be), and if it's not, the task is
skipped (and zabbix_server_selinux remains true).

Also drop checking os_family, we're only included from RedHat.yml,
which is only included for os_family RedHat.