Rework/Cleanup Agent Role (#1229)
This PR cleans up a lot of duplicative code within the Agent role. It standardizes on the use of the `zabbix_agent` prefix and deprecates the `zabbix_agent2` prefix to make the code more reusable. Documentation is cleaned up and the config file is moved to the same format as we have done for the other roles.