roles/agent: refactor detect_ip tasks
It seems the original code worked like this;

1. Always set to default_ipv4.address
2. if that doesn't work, check if we don't not detect ip
3. get a count of private ipv4 addresses
4. if the private ipv4 address count is 0, set it to the first public
5. else set to the first private.
with some additional checks to make this more confusing.

We can simplify it to;

1. Always set ip to default_ipv4.address if available.
2. If we need to detect ip
3a. Pick first private ipv4 address, if any available
3b. Pick first public ipv4 address, if any available

Also rename variable for agent_ip code,
`not (zabbix_agent_dont_detect_ip)` is just an unfortunate read, so
this change makes the when condition read better, whilst retaining
compatability.