roles/agent: refactor listeninterface
So this code was pretty useless. `zabbix_agent_listenip` being already
defined in `defaults/main.yml` most of the tasks are just skipped.
I've reinterpreted the tasks to make it atleast somewhat useful again;

If someone specifies zabbix_agent_listeninterface, we will verify the
interface exists, and then try to override listenip to that of the
interface.

Also use assert instead of fail, which makes it clearer *WHAT* we
are asserting and *WHEN*, and the task is marked as OK rather than
skipped.

But I feel this is pretty much dead code. If the user already knows
their listeninterface during deployment, they could just as easliy added
zabbix_agent_listenip: "{{ ansible_facts['eno1'].ipv4.address }}" to
their hostvars instead.