Agent refactor (part 1) (#1484)
* roles/agent: refactor detect_ip tasks

It seems the original code worked like this;

1. Always set to default_ipv4.address
2. if that doesn't work, check if we don't not detect ip
3. get a count of private ipv4 addresses
4. if the private ipv4 address count is 0, set it to the first public
5. else set to the first private.
with some additional checks to make this more confusing.

We can simplify it to;

1. Always set ip to default_ipv4.address if available.
2. If we need to detect ip
3a. Pick first private ipv4 address, if any available
3b. Pick first public ipv4 address, if any available

Also rename variable for agent_ip code,
`not (zabbix_agent_dont_detect_ip)` is just an unfortunate read, so
this change makes the when condition read better, whilst retaining
compatability.

* roles/agent: Move when clause before block

Part of ansible-lint best practices. when conditions have a tendency
to get lost while modifying tasks in a block.

* roles/agent: FIX exclude darwin/macos from repo

The zabbix_repo role fails, because we don't have vars/Darwin.yml
in that role. (And also because macos does things differently.)

* roles: declutter defaults by moving internal vars

Variables prefixed with underscore are internal vars not really
meant for overriding, and are better suited under vars/*yml

* roles: sort database defaults in uri order

* roles/repo: default to 7.2

The other roles already default to 7.2.

* roles/agent: refactor selinux

Let's simplify the crap out of this.

* molecule/agent: We're not testing against Sangoma

We don't have any containers that will report os_family Sangoma

* molecule/agent: Remove unwarranted packages

net-tools went out of style atleast two decades ago, sometime 2001.
Then someone kindly pointed that out again in 2009 [1]. You should
not be using this package, nor these tools.

The installation of these packages feel like some old relic for
someone wanting some debug tools inside the container, and not
really relevant for zabbix-agent itself.

[1] https://lists.debian.org/debian-devel/2009/03/msg00780.html

* molecule/agent: remove misplaced when condtion

This condition doesn't make sense, in regards to the other tasks
which are performed without similar conditions.

* molecule/agent: remove unused docker tasks

This code is RedHat specific and probably hasn't seen use in quite
some while. I fail to see how this even gets used during molecule
testing.

* roles/agent: move set_facts into defaults

Defaults allow you to provide a value, that's easily overridable by
the user (by having a lower precedence [1]). So we don't have to
check if a user overrides a value, before we set the default.

[1] https://docs.ansible.com/ansible/latest/playbook_guide/playbooks_variables.html#understanding-variable-precedence

* roles/agent: Refactor service

It's the same code for Debian, RedHat and Suse, so consolidate those
tasks into one. We flush the handlers so any changes to the
configuration that force a restart are run first. And then ensure
the zabbix-agent service is enabled and started.

* roles/agent: move docker tasks

These docker-specific tasks can run from within the Docker.yml

* roles/agent: refactor listeninterface

So this code was pretty useless. `zabbix_agent_listenip` being already
defined in `defaults/main.yml` most of the tasks are just skipped.

I think if someone specifies zabbix_agent_listeninterface, we were
supposed to verify the interface exists, and then try to override
listenip to that of the interface. But there's really no point in
doing this. If the user already knows their listeninterface during
deployment, they could just as easliy added

zabbix_agent_listenip: "{{ ansible_facts['eno1'].ipv4.address }}"

to their group/host vars.