roles/agent: minor docker cleanup
If zabbix_agent_docker we run the configure-Docker.yml playbook.

We create a system user and system group for zabbix

zabbix_agent_tlspskfile is always defined, so we need to check for
psk in tls_{accept,connect} instead.

Also, the docker environment needed some small modifications in how
the variables are presented to the container.