Agent refactor (#1485)
* roles/agent: install-{{ os_family }}.yml

Intermediate step to make future changes easier to follow.

Rename these task-files to install-{{ os_family }}, since that is
what's happening in them, more or less.

* roles/agent: refactor zabbix_agent_package

Put the override/calculation in vars/RedHat.yml, and default to
zabbix-agent(2) for the other distros.

* roles/agent: refactor zabbix-{get,sender}

Put the override/calculation in vars/RedHat.yml, and default to
zabbix-{get,sender} for the other distros.

* roles/agent: consolidate Debian, RedHat, and Suse installations

It's pretty much the same tasks for Debian, RedHat, and Suse. So
by using a with_first_found, we load os_family specific files if
they exist (Windows, Darwin), and then try the "generic" install
for the rest.

* * roles/agent: move set_facts into defaults

Defaults allow you to provide a value, that's easily overridable by
the user (by having a lower precedence [1]). So we don't have to
check if a user overrides a value, before we set the default.

This cleans up some of the variable use by Windows tasks.

[1] https://docs.ansible.com/ansible/latest/playbook_guide/playbooks_variables.html#understanding-variable-precedence

* roles/agent: bugfix _listen_ip => _listenip

* roles/agent: zabbix_agent2 is defined in defaults

* roles/agent: refactor configure step

Looks like these files contain most of the configuring of the
agents, so we use the same with_first_found construction as for
install.

* roles/agent: move psk configuration

The psk configuration tasks contain os_family specific tasks, so
lets move it into all os_family specific tasks so we can refactor
those things out.

* roles/agent: move os tasks out of psk_secret.yml

Now each os_family task can register directly to zabbix_agent_tlspskcheck
and we can drop the conditional set_fact task.

* roles/agent: additional os tasks out of psk_secret.yml

Looks like we might have accidentally broken Darwin/macOS in some
of the previous commits. We'll get back to fixing it eventually.

* roles/agent: move os tasks out of psk_identity.yml

Now each os_family task can register directly to zabbix_agent_tlspskidentity_check
and we can drop the conditional set_fact task.

* roles/agent: additional os tasks out of psk_identity.yml

Looks like we might have accidentally broken Darwin/macOS in some
of the previous commits. We'll get back to fixing it eventually.

* roles/agent: clean up psk_{secret,identity}.yml

We should be inheriting the config tag from the first block including
these files.

* roles/agent: refactor zabbix_agent_logfile

Might want to consider splitting this into directory and filename,
so we can reuse

  zabbix_agent2 | ternary('zabbix_agent2.log', 'zabbix_agentd.log')

But for now, just reduce the task(s) involved in setting this variable.

* roles/agent: refactor zabbix_agent_include_dir(s)

This simplifies the windows set_fact overrides, and just pushes
zabbix_agent_include_dir over the edge to become a list. For some
reason, there was a need to specify multiple directories, and this
commit finalizes that process.

* roles/agent: eliminate task setting controlsocket

This is better solved with defaults/variables/overrides.

* roles/agent: move zabbix_agent_ip tasks to api.yml

They're only relevant for when you're registering a host via the
api, so let's move them there.

* roles/agent: get rid of set_facts tasks

The less set_facts the better.

* roles/agent: move service to configure.yml

Technically it's already there, so remove the duplicate from main,
and move it to the very end, with flush_handlers.

* roles/agent: remove redundant docker volume mounts

If you're mounting / at hostfs/, then all these other mounts will
follow.

* roles/agent: default docker image-tag to latest

Also, use the zabbix_agent2 variable to set which agent to run.

* roles/agent: bugfix, correct volume-mounts

Both docker images expect the voulme-mount at
/etc/zabbix/zabbix_agentd.d, so it's supposed to be the other way
around.

* roles/agent: minor docker cleanup

If zabbix_agent_docker we run the configure-Docker.yml playbook.

We create a system user and system group for zabbix

zabbix_agent_tlspskfile is always defined, so we need to check for
psk in tls_{accept,connect} instead.

Also, the docker environment needed some small modifications in how
the variables are presented to the container.

* roles/agent: refactor windows install

This hasn't worked in some time, and could use some cleanup.

The whole install process has gotten simpler with the msi packages.
Instead of using the pinned zabbix_version_long to 5.2.4, we query
the latest zabbix releases to get the long version. And a bunch of
ternary operations to determine zabbix-agent vs zabbix-agent2
files/services/locations.

The whole uninstall agent needs some improvement, I'm leaving it in
incase someone more powershell savy care to take a look.