roles/repo: fix architecture for debian packages
Zabbix provides packages for many archtectures on debian;
i386, amd64, armhf, and arm64.

So we need a lookup/map for these. I'm switching the check from
machine to architecture in ansible_facts, because architecture is a
closer match to whatever the operatingsystem compiled for, rather
than the machine type it's running on. (machine can report i686,
while the architecture reports i386.)

This is not foolproof however as you can run an armhf based os on a
64-bit arm board, and ansible will report the architecture as
aarch64. (quite possibly an ansible bug)

So we put the zabbix_repo_deb_arch to defaults, so it can be
overridden by the user in such cases.