molecule/prepare: remove/clarify policy-rc.d task
I think we should drop this task completely (we're not using this
for zabbix_web), but for now just clarifying what it's for feels
like a small victory.

Dropping the apt-get update, as we can pass update_cache: true to
any package installation task and therefore no need to do this.

There should not really be any services installed or required to run
during our setup of the zabbix_repo, so just remove the task.