roles/javagw: remove service stopping task
I tried to go back in the logs to find a reason for this. We don't
do this for any other services, and from a security standpoint this
is not the way to do it. If you can't afford to have the service
running unconfigured for 2 minutes before the playbook configures
it and restarts the service, you can't afford to have it start up
at all. In that case we should be masking the service, or using
policy-rc.d or some other solution to prevent it from starting up
after install.