roles/javagw: refactor and consolidate tasks
We've done this for zabbix_{server,proxy} already, so a single task
for the installation all distributions.

Rule of thumb, if you're doing less than 5 tasks (and not looping
an include task), it probably doesn't deserve its own task file, so
the remaining os_family specific tasks can be merged in to main.yml

I'm a bit unsure about installing openjdk for Suse, why aren't we
installing openjdk for other distros?