Arm64 (#1495)
* roles/repo: fix debian/ubuntu arm64 repo url

Apparently they've simplified the url for zabbix 7.2, which is a
welcome change.

* roles/repo: set zabbix_repo_deb_url in defaults

Simplifies the playbook, and shows variables available for overrides
by users more clearly.

* roles/repo: move internal vars to vars/

Whenever we prefix a variable with _, we're essentially saying it's
an internal variable not meant to be overridden by users, so it's
cleaner if they're not presented in defaults.

* roles/repo: fix architecture for debian packages

Zabbix provides packages for many archtectures on debian;
i386, amd64, armhf, and arm64.

So we need a lookup/map for these. I'm switching the check from
machine to architecture in ansible_facts, because architecture is a
closer match to whatever the operatingsystem compiled for, rather
than the machine type it's running on. (machine can report i686,
while the architecture reports i386.)

This is not foolproof however as you can run an armhf based os on a
64-bit arm board, and ansible will report the architecture as
aarch64. (quite possibly an ansible bug)

So we put the zabbix_repo_deb_arch to defaults, so it can be
overridden by the user in such cases.

* changelog/fragment for bugfixes