Do not inject facts as vars (#1497)
* github/actions: remove duplicate excludes

* github/actions: remove testing for v64

Strategy jobs for molecule now exceed the maximum of 256
configurations. So let's remove testing for 6.4 which is no longer
supported by Zabbix.

* roles/repo: remove proxy hanlder (redhat)

I'm not sure when this was originally introduced, and what it's
for. It seems to want to remove the proxy setting from repo files,
which is counter-productive to those wanting to use a proxy.

* roles/repo: Set inject_facts_as_vars flase

* roles/server: replace set_fact with defaults

This type of set_fact usage is cleaner when supplied as defaults

* roles/server: Set inject_facts_as_vars flase

* molecule/server: Remove unwarranted packages

net-tools went out of style atleast two decades ago, sometime 2001.
Then someone kindly pointed that out again in 2009 [1]. You should
not be using this package, nor these tools.

The installation of these packages feel like some old relic for
someone wanting some debug tools inside the container, and not
really relevant for zabbix-server itself.

[1] https://lists.debian.org/debian-devel/2009/03/msg00780.html

* roles/proxy: Set inject_facts_as_vars flase

* molecule/proxy: Remove unwarranted packages

net-tools went out of style atleast two decades ago, sometime 2001.
Then someone kindly pointed that out again in 2009 [1]. You should
not be using this package, nor these tools.

The installation of these packages feel like some old relic for
someone wanting some debug tools inside the container, and not
really relevant for zabbix-proxy itself.

[1] https://lists.debian.org/debian-devel/2009/03/msg00780.html

* roles/web: Set inject_facts_as_vars flase

* molecule/web: geerlingguy workaround

* molecule/web: Remove unwarranted packages

net-tools went out of style atleast two decades ago, sometime 2001.
Then someone kindly pointed that out again in 2009 [1]. You should
not be using this package, nor these tools.

The installation of these packages feel like some old relic for
someone wanting some debug tools inside the container, and not
really relevant for zabbix-web itself.

[1] https://lists.debian.org/debian-devel/2009/03/msg00780.html

* molecule/prepare: remove/clarify policy-rc.d task

I think we should drop this task completely (we're not using this
for zabbix_web), but for now just clarifying what it's for feels
like a small victory.

Dropping the apt-get update, as we can pass update_cache: true to
any package installation task and therefore no need to do this.

There should not really be any services installed or required to run
during our setup of the zabbix_repo, so just remove the task.

* github/actions: use the generic "Run role tests"

They're not all testing the server role. Removing this copy-pasta.

* roles/javagw: Set inject_facts_as_vars flase

* roles/{server,proxy}: simplify timescale/sqlite3 import

Simplifying timescale and sqlite3 import, by reusing *_database_file_dir.

Moving '_' prefixed internal vars to vars/ to declutter defaults.

And eliminating the trailing slash from the variable makes the code
read easier with a natural / separating the variable and the path
continuation.

* roles/javagw: remove service stopping task

I tried to go back in the logs to find a reason for this. We don't
do this for any other services, and from a security standpoint this
is not the way to do it. If you can't afford to have the service
running unconfigured for 2 minutes before the playbook configures
it and restarts the service, you can't afford to have it start up
at all. In that case we should be masking the service, or using
policy-rc.d or some other solution to prevent it from starting up
after install.

* roles/javagw: refactor and consolidate tasks

We've done this for zabbix_{server,proxy} already, so a single task
for the installation all distributions.

Rule of thumb, if you're doing less than 5 tasks (and not looping
an include task), it probably doesn't deserve its own task file, so
the remaining os_family specific tasks can be merged in to main.yml

I'm a bit unsure about installing openjdk for Suse, why aren't we
installing openjdk for other distros?

* roles/javagw: remove zabbix 5.0 remnant

This (probably) has to do with where the sql-scripts were installed
under zabbix 5.0.

* molecule/javagw: drop apt-get update in workaround

We don't need this apt-get update, so lets use the file module to
create the missing directory.

* molecule/javagw: Remove unwarranted packages

net-tools went out of style atleast two decades ago, sometime 2001.
Then someone kindly pointed that out again in 2009 [1]. You should
not be using this package, nor these tools.

The installation of these packages feel like some old relic for
someone wanting some debug tools inside the container, and not
really relevant for zabbix-java-gateway itself.

[1] https://lists.debian.org/debian-devel/2009/03/msg00780.html

* github/actions: update pinned galaxy versions

Probably time to update these, see what shakes out.

* Add changelog fragment

* molecule/repo: Remove another remnant of 5.0