Sane Selinux defaults (#1537)
* roles/selinux: provide sane defaults

zabbix-server and zabbix-proxy are pretty useless without networking,
and the web interface keeps complaining if it can't connect to zabbix
over port 10050, so might aswell let it, and allow it to connect to
databases over the network, as we don't have configuration in place
for using unix sockets.

* roles/proxy: remove daemons_enable_cluster_mode

I don't think this is ever required by zabbix-proxy.

* selinux changelog/fragment