Sane defaults (#1608)
* roles/server: fix zabbix_repo_package
* roles/repo: default to 7.4

* roles/repo: default value for apt priority

Leaving this undefined is just bad form, so might aswell give it
a value. And ff we're actually taking the time to install repo
configuration for the official packages, it's a good indicator that
we also want the packages we install to come from that repository.

P >= 1000 forces these packages even if there are newer available
elsewhere. [1]

[1] https://manpages.debian.org/buster/apt/apt_preferences.5.en.html#How_APT_Interprets_Priorities

* roles/repo: enable gpg checking

Since the variable was misspelled, yum never performed GPG signature
checks. But since really should have sane defaults, we're now doing
those checks. And we need to provide the correct keys and some overrides
for v60 and EL8.