Add httptest module for Web Scenario
Adds support for httptest management only for host, due to the lack of
information in the documentation about the divergence of behaviour of the
API between host and template.

Some info can be found in changelogs like the following in the 3.0
changelog:
 - ZBX-11191 httptest.update: fixed updating templated web scenario steps
   by prohibiting to directly change the step name or giving different amount
   of steps than in template

But nothing is properly documented on 7.4 or even 8.0 documentations.