fix: zabbix-web configuration group
Zabbix-frontend-php packages for Debian distros perform a `chown` on the
`/etc/zabbix/web` directory with a `www-data:root` owner:group.
The `chown` operation is done in the package `postinst` script.

Having "{{ zabbix_php_fpm_conf_group }}" as the group owner of this directory
is overwritten at package update time, causing unnecessary changes in ansible.

This commit changes the owning group to `root` (gid 0) to fix the issue.